+++ app/Http/Filters/CocktailMethodQueryFilter.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Kami\Cocktail\Http\Filters;
 
+use Kami\Cocktail\Models\Cocktail;
 use Spatie\QueryBuilder\QueryBuilder;
 use Spatie\QueryBuilder\AllowedFilter;
 use Kami\Cocktail\Models\CocktailMethod;
@@ -22,7 +23,12 @@ final class CocktailMethodQueryFilter extends QueryBuilder
                 AllowedFilter::partial('name'),
             ])
             ->defaultSort('name')
-            ->withCount('cocktails')
+            ->addSelect([
+                'cocktails_count' => Cocktail::query()
+                    ->selectRaw('count(*)')
+                    ->whereColumn('cocktail_methods.id', 'cocktails.cocktail_method_id')
+                    ->where('cocktails.bar_id', bar()->id)
+            ])
             ->filterByBar();
     }
 }